# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Laravel",
        "contact": {
            "name": "hello@example.com",
            "email": "hello@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User profile",
                "description": "User profile",
                "operationId": "profile_show",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "User logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register new User",
                "operationId": "register",
                "requestBody": {
                    "description": "Register new User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string",
                                        "example": "Kasun Vimarshana"
                                    },
                                    "username": {
                                        "description": "Username",
                                        "type": "string",
                                        "example": "kasun"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User confirm password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "role": {
                                        "description": "User Role",
                                        "type": "string",
                                        "example": "owner|manager|cashier"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "JsonResponseResource": {
                "title": "JsonResponseResource",
                "description": "Json Response Resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status",
                        "type": "string"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data",
                        "nullable": true
                    },
                    "errors": {
                        "title": "Errors",
                        "description": "Errors",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonResponseResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}